+++ qiskit_ibm_runtime/decoders/executor_estimator/utils.py
@@ -14,6 +14,7 @@
 
 from __future__ import annotations
 
+from functools import lru_cache
 from typing import TYPE_CHECKING
 
 import numpy as np
@@ -160,3 +161,38 @@ def compute_exp_val(
 
     # Ensure we always return numpy arrays (even for scalar results)
     return np.asarray(exp_val), np.asarray(ensemble_variance), np.asarray(twirl_variance)
+
+
+def unbroadcast_index(
+    bc_index: tuple[int | slice, ...], shape: tuple[int, ...]
+) -> tuple[int | slice, ...]:
+    """Index an array using an index from a compatible broadcasted shape.
+
+    An ND-array ``arr`` is broadcastable to any shape ``bc_shape = (*pad_shape, *arr.shape)``.
+    This function allows indexing ``arr`` using an ND-index or slice from ``bc_shape`` and
+    will return the index for ``arr`` that accesses the same value.
+
+    Args:
+        bc_index: An ND-index from a broadcasted shape.
+        shape: The shape of the broadcasting compatible array to index.
+
+    Returns:
+        The equivalent un-broadcasted ND-index of the array with specified shape.
+    """
+
+    @lru_cache
+    def _pad_broadcast_shape(shape: tuple[int, ...], ndims: int) -> tuple[int | slice, ...]:
+        # Pad a shape with trivial dimensions.
+        shape_ndims = len(shape)
+        pad = ndims - shape_ndims
+        if pad > 0:
+            return pad * (1,) + shape
+        return shape
+
+    shape_ndims = len(shape)
+    if shape_ndims == 0:
+        return ()
+
+    pad_shape = _pad_broadcast_shape(shape, len(bc_index))
+    bc_index = tuple(0 if dim == 1 else i for i, dim in zip(bc_index, pad_shape))
+    return bc_index[-shape_ndims:]
